Feat/aasdk 490 close button metric common (#357)
* chore(): WIP

* feat(performance metrics): Added close button pressed time metric tracking for common

* chore(): cleanup

* chore(): cleanup

* chore(): additional tests

* chore(): cleanup

* chore(): cleanup

+++ superawesome-common/src/main/java/tv/superawesome/sdk/publisher/common/models/PerformanceMetric.kt
@@ -1,6 +1,6 @@
 package tv.superawesome.sdk.publisher.common.models
 
-internal data class PerformanceMetric(val value: Double,
+internal data class PerformanceMetric(val value: Long,
                                       val metricName: PerformanceMetricName,
                                       val metricType: PerformanceMetricType) {
   fun build(): Map<String, Any> {
